requirements.txt
kubernetes==28.1.0
prometheus-api-client==0.5.4
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
matplotlib==3.7.2
PyYAML==6.0.1